Prevent 'false' from showing up in the console report
Prior to this, when viewing the report in the console after diffing to
catalog files, the word "false" would show up right next to the number
of "Nodes with the most changes by differences".

For example, note the errant "false" at the end of this:
```
--------------------------------------------------------------------------------
1 out of 1 nodes changed.                                   0.34%
--------------------------------------------------------------------------------

Nodes with the most changes by percent changed:
1. foo.bar.net-production                                     0.34%

Nodes with the most changes by differences:
1. foo.bar.net-some_feature                                   8false
```

This was due to a Ruby `.key?` check inside a string interpolation of
`nodes[:puppet_ouput]`, which when ":pull_output" was not a key in the
`nodes` hash evaluated to `false`. And since `false` was inside the
string interpolation, it got printed to the console.

This fixes that by switching to the "x if condition y" format of
conditional logic, which hides the output of the `.key?` check.

After this commit, the console output does not contain the word "false".
```
--------------------------------------------------------------------------------
1 out of 1 nodes changed.                                   0.34%
--------------------------------------------------------------------------------

Nodes with the most changes by percent changed:
1. foo.bar.net-production                                     0.34%

Nodes with the most changes by differences:
1. foo.bar.net-some_feature                                   8
```